Reformat and move more stuff around
