+++ test_reqs.http
@@ -1,7 +1,6 @@
 ### Send a new resource
 POST http://localhost:8080/apis/apiextensions.k8s.io/v1/customresourcedefinitions
 Content-Type: application/json
-
 {
   "metadata": {
     "name": "foobar"
